+++ CHANGELOG.md
@@ -16,6 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Improved console output in verbose mode.
 * Improved logging from multiprocessing jobs.
 * Added PMDS layout algorithm.
+* Remove multi-sample processing from all `single-cell` subcommands
+* Add `--sample_name` option to `single-cell amplicon` to overwrite the name derived from the input filename.
+* Add `--skip-input-checks` option to `single-cell amplicon` to make input filename checks warnings instead of errors.
+
+### Removed
+
+* Remove `--input1_pattern` and `--input2_pattern` from `single-cell amplicon` command.
 
 ## [0.16.1] - 2024-01-12